Added POST and PUT for prices

+++ ToDoApi/Model/Prices.cs
@@ -5,12 +5,12 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string PriceCode { get; set; }
-        public float? Hood { get; set; }
-        public float? Gown { get; set; }
-        public float? Hat { get; set; }
-        public float? XtraHood { get; set; }    
-        public float? UcolSash { get; set; }
-        public float? Freight { get; set; }
+        public decimal? Hood { get; set; }
+        public decimal? Gown { get; set; }
+        public decimal? Hat { get; set; }
+        public decimal? XtraHood { get; set; }    
+        public decimal? UcolSash { get; set; }
+        public decimal? Freight { get; set; }
         public string? PriceNote { get; set; }
     }
 }
